Moved FeatureElementListener to JUnit listeners
