+++ src/Behat/Behat/Output/ServiceContainer/Formatter/JUnitFormatterFactory.php
@@ -96,7 +96,7 @@ class JUnitFormatterFactory implements FormatterFactory
         $definition = new Definition('Behat\Testwork\Output\Node\EventListener\ChainEventListener', array(
             array(
                 new Reference('output.node.listener.junit.outline'),
-                new Definition('Behat\Behat\Output\Node\EventListener\AST\FeatureElementListener', array(
+                new Definition('Behat\Behat\Output\Node\EventListener\JUnit\JUnitFeatureElementListener', array(
                     new Reference('output.node.printer.junit.feature'),
                     new Reference('output.node.printer.junit.scenario'),
                     new Reference('output.node.printer.junit.step'),
